Fine Calculation implemented

+++ Backend/src/main/java/com/librarywebapp/Backend/DTO/Request/CustomerAddDTO.java
@@ -13,5 +13,5 @@ public class CustomerAddDTO {
     private String address;
     private String password;
     private String phoneNumber;
-    private Float fine;
+    private double fine;
 }
