+++ Backend/src/main/java/com/librarywebapp/Backend/Model/Customer.java
@@ -23,5 +23,5 @@ public class Customer {
     private String address;
     private String password;
     private String phoneNumber;
-    private Float fine;
+    private double fine;
 }
